+++ components/training/eventBooking.tsx
@@ -3,10 +3,11 @@ import classNames from "classnames";
 import dayjs from "dayjs";
 import React, { FC } from "react";
 import { MdLocationOn } from "react-icons/md";
+import { Template, wrapFieldsWithMeta } from "tinacms";
 import { tinaField } from "tinacms/dist/react";
 import { CustomLink } from "../customLink";
 import { Container } from "../util/container";
-import { eventBookingBlock } from "./eventBooking.schema";
+import { TimePicker } from "./eventBooking.schema";
 import { EventBookingType, EventModel } from "./eventBookingType";
 export const isEmpty = (value) => {
   return (
@@ -283,3 +284,137 @@ const EventLocation = ({ event }: { event: Event }) => {
     </span>
   );
 };
+
+const gstTypeOptions = ["inc GST", "+ GST"];
+
+export const eventBookingBlock = {
+  eventBooking: "EventBooking",
+  eventDurationInDays: "eventDurationInDays",
+  price: "price",
+  discountPrice: "discountPrice",
+  discountNote: "discountNote",
+  suffix: "suffix",
+  eventList: {
+    value: "eventList",
+    city: "city",
+    date: "date",
+    endDate: "endDate",
+    bookingURL: "bookingURL",
+    location: "location",
+  },
+  gstText: "gstText",
+};
+
+export const eventBookingSchema: Template = {
+  name: eventBookingBlock.eventBooking,
+  label: "Events Booking",
+  ui: {
+    previewSrc: "/images/thumbs/tina/events-booking.jpg",
+  },
+  fields: [
+    {
+      type: "number",
+      label: "Duration (In Days)",
+      name: eventBookingBlock.eventDurationInDays,
+      required: true,
+    },
+    {
+      type: "number",
+      label: "Price",
+      name: eventBookingBlock.price,
+      required: true,
+    },
+    {
+      type: "number",
+      label: "Discount Price",
+      name: eventBookingBlock.discountPrice,
+    },
+    {
+      type: "string",
+      name: eventBookingBlock.gstText,
+      label: "Select GST Option",
+      options: gstTypeOptions,
+      required: true,
+    },
+
+    {
+      type: "string",
+      label: "Discount Note",
+      name: eventBookingBlock.discountNote,
+    },
+    {
+      type: "object",
+      label: "Event",
+      name: eventBookingBlock.eventList.value,
+      ui: {
+        itemProps: (item) => {
+          return { label: item?.city };
+        },
+      },
+      list: true,
+      fields: [
+        {
+          type: "string",
+          label: "City",
+          name: eventBookingBlock.eventList.city,
+        },
+        {
+          type: "datetime",
+          label: "Start Date",
+          name: eventBookingBlock.eventList.date,
+          ui: {
+            //@ts-expect-error - TODO: update when tina datetime picker is updated to support utc https://github.com/tinacms/tina.io/issues/2989
+            utc: true,
+            format: (value) => {
+              const val =
+                value && dateFormat.format(new Date(Date.parse(value)));
+              return val;
+            },
+          },
+        },
+        {
+          label: "Start Time",
+          type: "datetime",
+          ui: {
+            component: wrapFieldsWithMeta(({ input }) => (
+              <TimePicker defaultValue={"09:00"} input={input} />
+            )),
+            //@ts-expect-error - TODO: update when tina datetime picker is updated to support utc https://github.com/tinacms/tina.io/issues/2989
+            utc: true,
+          },
+          name: "startTime",
+        },
+        {
+          label: "End Time",
+          type: "datetime",
+          ui: {
+            component: wrapFieldsWithMeta(({ input }) => (
+              <TimePicker input={input} defaultValue={"17:00"} />
+            )),
+            //@ts-expect-error - TODO: update when tina datetime picker is updated to support utc https://github.com/tinacms/tina.io/issues/2989
+            utc: true,
+          },
+          name: "endTime",
+        },
+        {
+          type: "string",
+          label: "Booking URL",
+          name: eventBookingBlock.eventList.bookingURL,
+        },
+        {
+          type: "reference",
+          label: "Location",
+          name: eventBookingBlock.eventList.location,
+          collections: ["locations"],
+        },
+      ],
+    },
+  ],
+};
+
+const dateFormat = Intl.DateTimeFormat("en-AU", {
+  timeZone: "UTC",
+  day: "numeric",
+  year: "numeric",
+  month: "short",
+});
